+++ HISTORY.rst
@@ -5,7 +5,7 @@ History
 Unreleased
 ----------
 * Use ``argparse`` for all command line tools and make use of ``workflows`` transport 
-  argument injection. Minimum ``workflows`` version is now 2.13
+  argument injection. Minimum ``workflows`` version is now 2.14
 
 0.10.0 (2021-10-04)
 -------------------